\name{resamplediversity-package}
\alias{resamplediversity-package}
\alias{resamplediversity}
\docType{package}
\title{
Between-population comparison of population diversity indices. Includes genetic data for brown bears (\emph{Ursus arctos}) in Northern Dinaric Mountains.
}
\description{
The package contains functions and data required to repeat analyses in the manuscript "Skrbinšek T, Jelenčič M, Waits LP, Potočnik H, Kos I, Trontelj P(2012) Using a reference population yardstick to calibrate and compare genetic diversity reported in different studies: an example from the brown bear. Heredity, In press". The functions can also be used to apply the "Reference population" approach described in the mentioned manuscript to other species/populations.
}
\details{
\tabular{ll}{
Package: \tab resamplediversity\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2012-07-02\cr
License: \tab GPL (>=2)\cr
}
}
\author{
Tomaž Skrbinšek \email{tomaz.skrbinsek@gmail.com}
}

\note{See package vignette by typing \code{vignette("resamplediversity")}.}
\references{
Skrbinšek T, Jelenčič M, Waits LP, Potočnik H, Kos I, Trontelj P(2012) Using a reference population yardstick to calibrate and compare genetic diversity reported in different studies: an example from the brown bear. Heredity, In press.
}

\keyword{datasets}

